refactor: extract the shared Option type into its own module
The `Option = OptionText | OptionAvatar` family (text/avatar dropdown option)
was declared identically in both SInputDropdown.vue and SInputSelectSearch.vue,
and imported from those components across the lens filter inputs + LensInputSelect
(which also re-exported it). Move the single definition to
`lib/components/Option.ts` and import it everywhere; drop the component-level
re-export. No behavior change.
