+++ stories/components/SInputDropdown.02_Avatar_Options.story.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
-import SInputDropdown, { type OptionAvatar, type Size } from 'sefirot/components/SInputDropdown.vue'
+import { type OptionAvatar } from 'sefirot/components/Option'
+import SInputDropdown, { type Size } from 'sefirot/components/SInputDropdown.vue'
 import { ref } from 'vue'
 
 const title = 'Components / SInputDropdown / 02. Avatar Options'
